web/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain==0.0.267
anthropic==0.3.11
openai==0.28.1
aiofiles==23.1.0
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0
